+++ options/index.js
@@ -1,4 +1,4 @@
-function onLoad() {
+document.addEventListener("DOMContentLoaded", () => {
     const ddByUrlAndThumb = document.getElementById('ddByUrlAndThumb');
     const ddByUrlOnly = document.getElementById('ddByUrlOnly');
     const dctHash = document.getElementById('dctHash');
@@ -10,6 +10,7 @@ function onLoad() {
     const showHashValues = document.getElementById('showHashValues');
     const reset = document.getElementById('reset');
 
+    /** Enable or disable UI elements based on deduplication mode. */
     function updateThumbnailHashElementsDisabled() {
         const disabled = ddByUrlOnly.checked;
         dctHash.disabled = disabled;
@@ -19,6 +20,25 @@ function onLoad() {
         partitionByDomain.disabled = disabled;
         showHashValues.disabled = disabled;
     };
+
+    /** Update UI elements according to the given settings. */
+    function updateUi(settings) {
+        if (settings.deduplicateThumbs) {
+            ddByUrlAndThumb.checked = true;
+        } else {
+            ddByUrlOnly.checked = true;
+        }
+        if (settings.hashFunction === 'diffHash') {
+            diffHash.checked = true;
+        } else {
+            dctHash.checked = true;
+        }
+        maxHammingDistance.valueAsNumber = settings.maxHammingDistance;
+        maxHammingDistanceText.value = maxHammingDistance.value;
+        partitionByDomain.checked = !!settings.partitionByDomain;
+        showHashValues.checked = !!settings.showHashValues;
+        updateThumbnailHashElementsDisabled();
+    }
 
     ddByUrlAndThumb.addEventListener('click', (event) => {
         browser.storage.local.set({deduplicateThumbs: true});
@@ -52,46 +72,10 @@ function onLoad() {
     });
 
     reset.addEventListener('click', (event) => {
-        ddByUrlAndThumb.checked = true;
-        dctHash.checked = true;
-        maxHammingDistance.valueAsNumber = 8;
-        maxHammingDistanceText.value = maxHammingDistance.value;
-        partitionByDomain.checked = true;
-        showHashValues.checked = false;
-        updateThumbnailHashElementsDisabled();
+        updateUi(defaultSettings);
         browser.storage.local.clear();
     });
 
-    async function restoreSettings() {
-        const settings = await browser.storage.local.get(
-            ['deduplicateThumbs', 'hashFunction', 'maxHammingDistance',
-             'partitionByDomain', 'showHashValues']);
-
-        if (settings.deduplicateThumbs === true) {
-            ddByUrlAndThumb.checked = true;
-        } else if (settings.deduplicateThumbs === false) {
-            ddByUrlOnly.checked = true;
-        }
-        if (settings.hashFunction === 'dctHash') {
-            dctHash.checked = true;
-        } else if (settings.hashFunction === 'diffHash') {
-            diffHash.checked = true;
-        }
-        if (settings.maxHammingDistance !== undefined) {
-            maxHammingDistance.valueAsNumber = settings.maxHammingDistance;
-            maxHammingDistanceText.value = settings.maxHammingDistance;
-        }
-        if (settings.partitionByDomain !== undefined) {
-            partitionByDomain.checked = !!settings.partitionByDomain;
-        }
-        if (settings.showHashValues !== undefined) {
-            showHashValues.checked = !!settings.showHashValues;
-        }
-        updateThumbnailHashElementsDisabled();
-    }
-
-    restoreSettings().catch(
-        (error) => console.warn("Failed to retrieve settings", error));
-}
-
-document.addEventListener("DOMContentLoaded", onLoad);
+    getSettings().then(updateUi).catch(
+        (error) => console.warn("Failed to restore settings", error));
+});